+++ src/main/scala/org/podval/docbook/gradle/PrepareDocBookTask.scala
@@ -1,6 +1,6 @@
 package org.podval.docbook.gradle
 
-import org.gradle.api.artifacts.Configuration
+import org.gradle.api.artifacts.{Configuration, Dependency}
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.{CopySpec, FileCopyDetails, RelativePath}
 import org.gradle.api.provider.{ListProperty, MapProperty, Property}
@@ -15,6 +15,12 @@ class PrepareDocBookTask extends DefaultTask  {
   private val layout: Layout = Layout.forProject(getProject)
 
   private val logger: Logger = new Logger.PluginLogger(getLogger)
+
+  @BeanProperty val xslt1version: Property[String] =
+    getProject.getObjects.property(classOf[String])
+
+  @BeanProperty val xslt2version: Property[String] =
+    getProject.getObjects.property(classOf[String])
 
   @Input @BeanProperty val inputFileName: Property[String] =
     getProject.getObjects.property(classOf[String])
@@ -33,21 +39,46 @@ class PrepareDocBookTask extends DefaultTask  {
 
   @TaskAction
   def prepareDocBook(): Unit = {
+    // Verify parameter sections
+    val allParameters: Map[String, Map[String, String]] =
+      parameters.get.asScala.toMap.mapValues(_.asScala.toMap)
+
+    val sectionsPresent: Set[String] = allParameters.keySet
+    val sectionsClaimed: Set[String] = DocBook2.processors.toSet[DocBook2].flatMap(_.parameterSections)
+    val sectionsUnclaimed: Set[String] = sectionsPresent -- sectionsClaimed
+    if (sectionsUnclaimed.nonEmpty) {
+      val sections: String = DocBook2.processors.map { processor =>
+        "  " + processor.name + ": " + processor.parameterSections.mkString(", ")
+      }.mkString("\n")
+
+      throw new IllegalArgumentException(
+        s"""Unsupported parameter sections: ${sectionsUnclaimed.mkString(", ")}.
+           |Supported sections are:
+           |$sections
+           |""".stripMargin
+      )
+    }
+
     // Input file
     writeInto(layout.inputFile(inputFileName.get), logger, replace = false) {
       """|<?xml version="1.0" encoding="UTF-8"?>
-         |<!DOCTYPE article PUBLIC "-//OASIS//DTD DocBook XML V5.0//EN" "http://www.oasis-open.org/docbook/xml/5.0/dtd/docbook.dtd">
-         |<article xmlns="http://docbook.org/ns/docbook" version="5.0" xmlns:xi="http://www.w3.org/2001/XInclude">
+         |<!DOCTYPE article
+         |  PUBLIC "-//OASIS//DTD DocBook XML V5.0//EN"
+         |  "http://www.oasis-open.org/docbook/xml/5.0/dtd/docbook.dtd">
+         |
+         |<article xmlns="http://docbook.org/ns/docbook" version="5.0"
+         |         xmlns:xi="http://www.w3.org/2001/XInclude">
          |</article>
          |"""
     }
 
     // XSLT stylesheets
-    unpackDocBookXsl(layout.forXslt1, "XSLT")
-    unpackDocBookXsl(layout.forXslt2, "XSLT 2.0")
+    unpackDocBookXsl(Stylesheets.xslt1, xslt1version.get)
+    unpackDocBookXsl(Stylesheets.xslt2, xslt2version.get)
 
     writeInto(layout.cssFile(cssFileName.get), logger, replace = false) {
-      """@namespace xml "http://www.w3.org/XML/1998/namespace";"""
+      """@namespace xml "http://www.w3.org/XML/1998/namespace";
+        |"""
     }
 
     // FOP configuration
@@ -68,29 +99,19 @@ class PrepareDocBookTask extends DefaultTask  {
 
     // Stylesheet files and customizations
 
-    val allParameters: Map[String, Map[String, String]] =
-      parameters.get.asScala.toMap.mapValues(_.asScala.toMap)
-
     val epubEmbeddedFontsStr: String =
       Fop.getFontFiles(layout.fopConfigurationFile, epubEmbeddedFonts.get.asScala.toList, logger)
 
-    (DocBook2.forXslt1 ++ DocBook2.forXslt2).foreach { processor: DocBook2 =>
-      // Parameters for this processor
-      val parameters: Map[String, String] = List(
-        allParameters.get("common"),
-        if (!processor.usesHtml) None else allParameters.get("common-html"),
-        allParameters.get(processor.name),
-        allParameters.get(processor.stylesheetName)
-      ).flatten.flatten.toMap
-
+    DocBook2.processors.foreach { processor: DocBook2 =>
       processor.writeStylesheetFiles(
-      layout = layout,
-      inputFileName = inputFileName.get,
-      parameters = parameters,
-      cssFileName = cssFileName.get,
-      epubEmbeddedFonts = epubEmbeddedFontsStr,
-      logger = logger
-    )}
+        layout = layout,
+        inputFileName = inputFileName.get,
+        parameters = processor.parameterSections.flatMap(allParameters.get).flatten.toMap,
+        cssFileName = cssFileName.get,
+        epubEmbeddedFonts = epubEmbeddedFontsStr,
+        logger = logger
+      )
+    }
 
     // substitutions DTD
     writeInto(layout.xmlFile(layout.substitutionsDtdFileName), logger, replace = true) {
@@ -120,11 +141,11 @@ class PrepareDocBookTask extends DefaultTask  {
          |
          |    <!-- DocBook XSLT 1.0 stylesheets  -->
          |    <rewriteURI uriStartString="http://docbook.sourceforge.net/release/xsl-ns/current/"
-         |                rewritePrefix="${layout.forXslt1.docBookXslDirectoryRelative}"/>
+         |                rewritePrefix="${layout.docBookXslDirectoryRelative(Stylesheets.xslt1.directoryName)}"/>
          |
          |    <!-- DocBook XSLT 2.0 stylesheets  -->
          |    <rewriteURI uriStartString="https://cdn.docbook.org/release/latest/xslt/"
-         |                rewritePrefix="${layout.forXslt2.docBookXslDirectoryRelative}"/>
+         |                rewritePrefix="${layout.docBookXslDirectoryRelative(Stylesheets.xslt2.directoryName)}"/>
          |
          |    <!-- generated data -->
          |    <rewriteSystem rewritePrefix="$data" systemIdStartString="data:/"/>
@@ -158,34 +179,41 @@ class PrepareDocBookTask extends DefaultTask  {
     }
   }
 
-  private def copyResourceInto(
-    directory: String,
-    name: String,
-    file: File
-  ): Unit = writeInto(file, logger, replace = false) {
-    Util.readFrom(getClass, s"/$directory/$name")
+  private def unpackDocBookXsl(stylesheets: Stylesheets, version: String): Unit = {
+    val directory: File = layout.docBookXslDirectory(stylesheets.directoryName)
+    if (!directory.exists) {
+      val dependencyNotation: String = stylesheets.dependencyNotation(version)
+      logger.info(s"Retrieving DocBook ${stylesheets.name} stylesheets: $dependencyNotation")
+      val file = getArtifact(stylesheets, dependencyNotation)
+      logger.info(s"Unpacking ${file.getName}")
+      unpack(
+        zipFile = file,
+        archiveSubdirectoryName = stylesheets.archiveSubdirectoryName,
+        directory = directory
+      )
+    }
   }
 
-  private def unpackDocBookXsl(layout: ForXslt, name: String): Unit = {
-    val directory: File = layout.docBookXslDirectory
-    if (!directory.exists) {
-      logger.info(s"Preparing DocBook $name stylesheets")
+  private def getArtifact(stylesheets: Stylesheets, dependencyNotation: String): File = {
+    val dependency: Dependency = getProject.getDependencies.create(dependencyNotation)
+    val configuration: Configuration = getProject.getConfigurations.detachedConfiguration(dependency)
+    configuration.setTransitive(false)
+    configuration.getSingleFile
+  }
 
-      val configuration: Configuration = getProject.getConfigurations.findByName(layout.docBookXslConfigurationName)
-      val zip: File = configuration.getSingleFile
-      val toDrop: Int = layout.docBookXslArchiveSubdirectoryName.count(_ == '/') + 1
-      getProject.copy((copySpec: CopySpec) => copySpec
-        .into(directory)
-        .from(getProject.zipTree(zip))
-        // following 7 lines of code deal with extracting just the "docbook" directory;
-        // this should become easier in Gradle 5.3, see:
-        // https://github.com/gradle/gradle/issues/1108
-        // https://github.com/gradle/gradle/pull/5405
-        .include(layout.docBookXslArchiveSubdirectoryName + "/**")
-        .eachFile((file: FileCopyDetails) =>
-          file.setRelativePath(new RelativePath(true, file.getRelativePath.getSegments.drop(toDrop): _*))
-        )
-        .setIncludeEmptyDirs(false))
-    }
+  private def unpack(zipFile: File, archiveSubdirectoryName: String, directory: File): Unit = {
+    val toDrop: Int = archiveSubdirectoryName.count(_ == '/') + 1
+    getProject.copy((copySpec: CopySpec) => copySpec
+      .into(directory)
+      .from(getProject.zipTree(zipFile))
+      // following code deals with extracting just the "docbook" directory;
+      // this should become easier in Gradle 5.3, see:
+      // https://github.com/gradle/gradle/issues/1108
+      // https://github.com/gradle/gradle/pull/5405
+      .include(archiveSubdirectoryName + "/**")
+      .eachFile((file: FileCopyDetails) =>
+        file.setRelativePath(new RelativePath(true, file.getRelativePath.getSegments.drop(toDrop): _*))
+      )
+      .setIncludeEmptyDirs(false))
   }
 }