+++ src/main/scala/org/podval/docbook/gradle/ProcessDocBookTask.scala
@@ -1,6 +1,6 @@
 package org.podval.docbook.gradle
 
-import org.gradle.api.{DefaultTask, Project}
+import org.gradle.api.DefaultTask
 import org.gradle.api.provider.{ListProperty, MapProperty, Property}
 import org.gradle.api.tasks.{Input, Internal, TaskAction}
 import java.io.File
@@ -14,8 +14,7 @@ class ProcessDocBookTask extends DefaultTask {
   private val logger: Logger = new Logger.PluginLogger(getLogger)
 
   // To let projects that use the plugin to not make assumptions about directory names:
-  @Internal def getOutputDirectory: File = layout.forXslt1.outputDirectoryRoot
-  @Internal def getOutputDirectory2: File = layout.forXslt2.outputDirectoryRoot
+  @Internal def getOutputDirectory: File = layout.outputDirectoryRoot
 
   // Register inputs
   val inputDirectories: Set[File] = Set(
@@ -23,10 +22,9 @@ class ProcessDocBookTask extends DefaultTask {
     layout.cssDirectory,
     layout.fopConfigurationDirectory,
     layout.dataDirectory,
-    layout.forXslt1.stylesheetDirectory
+    layout.stylesheetDirectory
   ) ++ Set(
     layout.imagesDirectory,
-    layout.forXslt2.stylesheetDirectory
   ).filter(_.exists)
 
   inputDirectories.foreach { directory: File =>
@@ -37,10 +35,8 @@ class ProcessDocBookTask extends DefaultTask {
 
   // Register outputs
   Set(
-    layout.forXslt1.saxonOutputDirectoryRoot,
-    layout.forXslt1.outputDirectoryRoot,
-    layout.forXslt2.saxonOutputDirectoryRoot,
-    layout.forXslt2.outputDirectoryRoot
+    layout.saxonOutputDirectoryRoot,
+    layout.outputDirectoryRoot
   ).foreach { directory: File =>
     Util.deleteRecursively(directory)
     logger.info(s"Registering output directory $directory")
@@ -59,24 +55,18 @@ class ProcessDocBookTask extends DefaultTask {
   @Input @BeanProperty val outputFormats: ListProperty[String] =
     getProject.getObjects.listProperty(classOf[String])
 
-  @Input @BeanProperty val outputFormats2: ListProperty[String] =
-    getProject.getObjects.listProperty(classOf[String])
-
   @Input @BeanProperty val isJEuclidEnabled: Property[Boolean] =
     getProject.getObjects.property(classOf[Boolean])
 
   @TaskAction
   def processDocBook(): Unit = {
     val processors: List[DocBook2] =
-      getProcessors("docBook.outputFormats", outputFormats, DocBook2.forXslt1, getProject)
-
-    val processors2: List[DocBook2] =
-      getProcessors("docBook.outputFormats2", outputFormats2, DocBook2.forXslt2, getProject)
+      Option(getProject.findProperty("docBook.outputFormats"))
+        .map(_.toString.split(",").map(_.trim).toList.filter(_.nonEmpty))
+        .getOrElse(outputFormats.get.asScala.toList)
+        .map(forName)
 
     logger.info(s"Output formats: ${getNames(processors)}")
-
-    if (processors2.nonEmpty)
-      logger.info(s"Output formats using XSLT 2.0: ${getNames(processors2)}")
 
     // In processing instructions and CSS, substitute xslParameters also - because why not?
     val allSubstitutions: Map[String, String] =
@@ -85,7 +75,7 @@ class ProcessDocBookTask extends DefaultTask {
 
     val resolver: Resolver = new Resolver(layout.catalogFile,  logger)
 
-    (processors ++ processors2).foreach { _.run(
+    processors.foreach { _.run(
       layout = layout,
       isJEuclidEnabled = isJEuclidEnabled.get,
       inputFileName = inputFileName.get,
@@ -96,25 +86,14 @@ class ProcessDocBookTask extends DefaultTask {
     )}
   }
 
-  private def getProcessors(
-     propertyName: String,
-     property: ListProperty[String],
-     supported: List[DocBook2],
-     project: Project,
-  ): List[DocBook2] = {
-    def forName(name: String): DocBook2 = {
-      supported.find(processor => processor.name.equalsIgnoreCase(name)).getOrElse {
-        throw new IllegalArgumentException(
-          s"""Unsupported output format $name;
-             |  supported formats are: ${getNames(supported)}""".stripMargin
-        )
-      }
+  private def forName(name: String): DocBook2 = {
+    val supported: List[DocBook2] = DocBook2.processors
+    supported.find(processor => processor.name.equalsIgnoreCase(name)).getOrElse {
+      throw new IllegalArgumentException(
+        s"""Unsupported output format $name;
+           |  supported formats are: ${getNames(supported)}""".stripMargin
+      )
     }
-
-    Option(project.findProperty(propertyName))
-      .map(_.toString.split(",").map(_.trim).toList.filter(_.nonEmpty))
-      .getOrElse(property.get.asScala.toList)
-      .map(forName)
   }
 
   private def getNames(processors: List[DocBook2]): String =